LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-mute-button-rtl-positioning.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
</head>
<body>
<script type="text/javascript">

description("Testing that fullscreen media controls mute button does not overlap with other buttons in RTL mode.");

window.jsTestIsAsync = true;

const mediaControls = new MacOSFullscreenMediaControls({ width: 458, height: 100 });

mediaControls.element.style.position = "relative";
document.body.appendChild(mediaControls.element);

step();

function step()
{
    debug("");
    debug("Testing RTL mode with new controls");
    debug("usesLTRUserInterfaceLayoutDirection = false");
    mediaControls.usesLTRUserInterfaceLayoutDirection = false;
    
    setTimeout(() => {
        const container = mediaControls.element.querySelector('.buttons-container.left');
        const muteButton = container.querySelector('.mute.bar');
        
        if (!muteButton) {
            testFailed("Could not find mute button with class 'mute.bar'");
            done();
            return;
        }
        
        const muteComputedStyle = window.getComputedStyle(muteButton);
        const leftValue = muteComputedStyle.getPropertyValue('left');
        const positionValue = muteComputedStyle.getPropertyValue('position');

        const volumeSlider = container.querySelector('.slider, .volume');
        if (volumeSlider) {
            const muteRect = muteButton.getBoundingClientRect();
            const sliderRect = volumeSlider.getBoundingClientRect();
            
            debug("");
            debug("Element positions in RTL:");
            debug(`  Mute button: left=${muteRect.left.toFixed(2)}, right=${muteRect.right.toFixed(2)}, width=${muteRect.width.toFixed(2)}`);
            debug(`  Volume slider: left=${sliderRect.left.toFixed(2)}, right=${sliderRect.right.toFixed(2)}, width=${sliderRect.width.toFixed(2)}`);

            const hasOverlap = muteRect.left < sliderRect.right && muteRect.right > sliderRect.left;
            if (!hasOverlap) {
                testPassed("Mute button and volume slider do not overlap in RTL mode");
            } else {
                testFailed("Mute button and volume slider should not overlap in RTL mode");
            }

            if (muteRect.left >= sliderRect.right) {
                testPassed("Mute button is correctly positioned to the right of slider in RTL mode");
            } else if (muteRect.right <= sliderRect.left) {
                testPassed("Mute button is correctly positioned to the left of slider in RTL mode");
            } else {
                testFailed(`Elements are overlapping or incorrectly positioned`);
            }
        }

        const centerContainer = mediaControls.element.querySelector('.buttons-container.center');
        if (centerContainer) {
            const playPauseButton = centerContainer.querySelector('.play-pause, button[id*="play"]');
            if (playPauseButton) {
                const muteRect = muteButton.getBoundingClientRect();
                const playPauseRect = playPauseButton.getBoundingClientRect();
                
                const hasOverlapWithPlayPause = muteRect.left < playPauseRect.right && muteRect.right > playPauseRect.left;
                if (!hasOverlapWithPlayPause) {
                    testPassed("Mute button does not overlap with play/pause button in RTL mode");
                } else {
                    testFailed("Mute button should not overlap with play/pause button in RTL mode");
                }
            }
        }
        
        done();
    }, 100);
}

function done()
{
    debug("");
    mediaControls.element.remove();
    finishJSTest();
}

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
